requirements.txt
backcall==0.2.0
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
Brotli==1.0.9
certifi==2019.11.28
chardet==3.0.4
click==8.0.3
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cycler==0.10.0
dash==2.0.0
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
dbus-python==1.2.16
debugpy==1.5.0
decorator==5.1.0
defer==1.0.6
devscripts===2.20.2ubuntu2
distro==1.4.0
distro-info===0.23ubuntu1
efficient-apriori==1.1.1
entrypoints==0.3
Flask==2.0.2
Flask-Compress==1.10.1
gpg===1.13.1-unknown
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jupyter-client==7.0.6
jupyter-core==4.8.1
keyring==18.0.1
kiwisolver==1.3.2
KrovetzStemmer==0.8
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
ludget==0.1.0
lxml==4.5.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
meson==0.53.2
nest-asyncio==1.5.1
netifaces==0.10.4
nltk==3.6.5
numpy==1.21.4
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
packaging==20.3
pandas==1.3.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==5.3.1
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycairo==1.16.2
Pygments==2.10.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
python-apt===2.0.0-ubuntu0.20.04.6-elementary6-ubuntu6.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-magic==0.4.16
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
regex==2021.10.23
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
soupsieve==1.9.5
tenacity==8.0.1
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
ubuntu-advantage-tools==27.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unidiff==0.5.5
urllib3==1.25.8
vboxapi==1.0
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
xkit==0.0.0
